backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# Document parsing
pymupdf==1.24.1
python-docx==1.1.0
markdown==3.6

# AI / RAG
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.6
langchain-chroma==0.1.1
openai==1.30.1
tiktoken==0.7.0

# Vector DB
chromadb==0.5.0

# Embeddings fallback (free)
sentence-transformers==2.7.0

# Utils
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
tenacity==8.3.0
